Fix some compilation warnings.
* buckets/hpack_buckets.c (handle_read_entry_and_clear):
   The precision argument for printf format "%.*" must be an int.

* buckets/log_wrapper_buckets.c
  (serf_log_wrapped_readline,
   serf_log_wrapped_read_iovec,
   serf_log_wrapped_read): Likewise.

* buckets/ssl_buckets.c
  (ssl_decrypt, ssl_encrypt): Likelwise.

* buckets/mmap_buckets.c
  (serf_mmap_read,
   serf_mmap_readline,
   serf_mmap_peek): Fix pointer types in output arguments.


git-svn-id: https://svn.apache.org/repos/asf/serf/trunk@1774378 13f79535-47bb-0310-9956-ffa450edef68

+++ buckets/hpack_buckets.c
@@ -1402,7 +1402,7 @@ handle_read_entry_and_clear(serf_hpack_decode_ctx_t *ctx,
 
     serf__log(LOGLVL_INFO, SERF_LOGCOMP_PROTOCOL, __FILE__, ctx->config,
               "Parsed from HPACK: %.*s: %.*s\n",
-              ctx->key_size, ctx->key, ctx->val_size, ctx->val);
+              (int)ctx->key_size, ctx->key, (int)ctx->val_size, ctx->val);
 
     if (!ctx->headers)
     {
